backend/requirements.txt
# TomeHub Backend Dependencies
# Production-ready requirements for Docker deployment

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Database
oracledb>=2.0.0

# AI & Embeddings
google-generativeai>=0.5.0

# Oracle Cloud Infrastructure
oci>=2.100.0

# Text Processing
rapidfuzz>=3.5.0
unidecode>=1.3.0

# PDF Processing
PyPDF2>=3.0.0

# EPUB Processing
ebooklib>=0.18
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Environment
python-dotenv>=1.0.0

# Production WSGI Server
gunicorn>=21.0.0

# Search & Ranking
rank_bm25>=0.2.2
python-json-logger
zeyrek>=0.1.0
symspellpy>=6.7.0
